+++ Scripts/Events/ModNanitesAdditionEvent.cs
@@ -6,6 +6,6 @@ namespace EventCallback
     public class ModNanitesAdditionEvent : Event<ModNanitesAdditionEvent>
     {
         //The amount that the nanite addition must be modified with, either subtracted or added
-        public int amount;
+        public float amount;
     }
 }
